requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.97.0
anyio==4.13.0
asgiref==3.11.1
billiard==4.2.4
celery==5.4.0
certifi==2026.4.22
click==8.3.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
distro==1.9.0
dj-database-url==3.1.2
Django==5.1
django-htmx==1.27.0
docstring_parser==0.18.0
factory_boy==3.3.3
Faker==40.15.0
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
iniconfig==2.3.0
jiter==0.14.0
kombu==5.6.2
packaging==26.2
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.12
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
PyJWT==2.12.1
pytest==9.0.3
pytest-asyncio==1.3.0
pytest-django==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==5.3.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
vine==5.1.0
wcwidth==0.6.0
whitenoise==6.12.0
